feat(relation): Introduce relation module
Remove eqnsimp. This is better be implemented as method than singledispatch.
